Rename `MutexLock` to `WorkspaceLock`
