+++ packages/langium/src/services.ts
@@ -35,7 +35,6 @@ import type { ScopeProvider } from './references/scope-provider.js';
 import type { JsonSerializer } from './serializer/json-serializer.js';
 import type { ServiceRegistry } from './service-registry.js';
 import type { AstReflection } from './syntax-tree.js';
-import type { MutexLock } from './utils/promise-util.js';
 import type { DocumentValidator } from './validation/document-validator.js';
 import type { ValidationRegistry } from './validation/validation-registry.js';
 import type { AstNodeDescriptionProvider, ReferenceDescriptionProvider } from './workspace/ast-descriptions.js';
@@ -64,6 +63,7 @@ import type { NodeKindProvider } from './lsp/node-kind-provider.js';
 import type { FuzzyMatcher } from './lsp/fuzzy-matcher.js';
 import type { IParserErrorMessageProvider } from 'chevrotain';
 import type { TypeHierarchyProvider } from './lsp/type-hierarchy-provider.js';
+import type { WorkspaceLock } from './workspace/workspace-lock.js';
 
 /**
  * The services generated by `langium-cli` for a single language. These are derived from the
@@ -179,8 +179,8 @@ export type LangiumDefaultSharedServices = {
         IndexManager: IndexManager
         LangiumDocuments: LangiumDocuments
         LangiumDocumentFactory: LangiumDocumentFactory
-        MutexLock: MutexLock
         TextDocuments: TextDocuments<TextDocument>
+        WorkspaceLock: WorkspaceLock
         WorkspaceManager: WorkspaceManager
     }
 }
